src/requirements/base.txt
fastapi==0.103.1
uvicorn==0.23.2
mypy==1.5.1
SQLAlchemy==2.0.20
PyMySQL==1.1.0
aiomysql==0.2.0
black==23.9.1
pydantic-settings==2.0.3
email-validator==2.0.0.post2
passlib==1.7.4
python-jose==3.3.0
sqlalchemy-stubs==0.4
python-multipart==0.0.6